Arrange VSIX unit test directory like product code directory
