+++ src/Sarif.Viewer.VisualStudio.UnitTests/Converters/ImportanceToBackgroundConverterTests.cs
@@ -1,42 +1,42 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Drawing;
 using System.Globalization;
 using FluentAssertions;
 using Microsoft.CodeAnalysis.Sarif;
 using Microsoft.Sarif.Viewer.Converters;
 using Xunit;
 
-namespace Microsoft.Sarif.Viewer.VisualStudio.UnitTests
+namespace Microsoft.Sarif.Viewer.VisualStudio.Converters.UnitTests
 {
-    public class ImportanceToForegroundConverterTests
+    public class ImportanceToBackgroundConverterTests
     {
         [Fact]
-        public void ImportanceToForegroundConverterHandlesUnimportant()
+        public void ImportanceToBackgroundConverterHandlesUnimportant()
         {
-            VerifyConversion(AnnotatedCodeLocationImportance.Unimportant, "Gray");
+            VerifyConversion(AnnotatedCodeLocationImportance.Unimportant, "Transparent");
         }
 
         [Fact]
-        public void ImportanceToForegroundConverterHandlesImportant()
+        public void ImportanceToBackgroundConverterHandlesImportant()
         {
-            VerifyConversion(AnnotatedCodeLocationImportance.Important, "Black");
+            VerifyConversion(AnnotatedCodeLocationImportance.Important, "Yellow");
         }
 
         [Fact]
-        public void ImportanceToForegroundConverterHandlesEssential()
+        public void ImportanceToBackgroundConverterHandlesEssential()
         {
-            VerifyConversion(AnnotatedCodeLocationImportance.Essential, "Black");
+            VerifyConversion(AnnotatedCodeLocationImportance.Essential, "Yellow");
         }
 
         private static void VerifyConversion(AnnotatedCodeLocationImportance importance, string expectedColor)
         {
-            var converter = new ImportanceToForegroundConverter();
+            var converter = new ImportanceToBackgroundConverter();
 
             string color = (string)converter.Convert(importance, typeof(string), null, CultureInfo.CurrentCulture);
 
             color.Should().Be(expectedColor);
         }
+
     }
 }
